+++ RELEASE-NOTES.md
@@ -41,6 +41,7 @@
 * Add `snow spcs service metrics` command to fetch service metrics:
   * Supports filtering by service name, container name, instance ID, and time intervals (`--since`, `--until`).
   * Use `--all` to fetch all columns.
+* Added `depends_on` field for entities in `snowflake.yml` that allows to specify order of calling the entites
 
 ## Fixes and improvements
 * Fixed inability to add patches to lowercase quoted versions